+++ src/components/LinkList/styles.module.css
@@ -1,0 +1,35 @@
+.linkList {
+  margin: 0;
+  padding: 0;
+  font-family: 'Karla', sans-serif;
+  list-style-type: none;
+  display: flex;
+}
+
+.linkList > li {
+  padding-bottom: 1rem;
+  padding-right: 1rem;
+}
+
+.link {
+  text-decoration: none;
+  background-repeat: no-repeat;
+  background-image: linear-gradient(150deg, #c3cfe2 0%, #f5f7fa 100%);
+  background-image: linear-gradient(-20deg, #b721ff 0%, #21d4fd 100%);
+  background-size: 100% 0.3rem;
+  background-position: 0 110%;
+  transition: background-size 100ms cubic-bezier(0.5, 0, 0.1, 1);
+  color: currentColor;
+}
+
+.link:hover,
+.link:focus {
+  background-size: 100% 100%;
+  background-image: linear-gradient(180deg, #f5f7fa 0%, #c3cfe2 100%);
+  background-image: linear-gradient(-20deg, #d887fe 0%, #78e6fe 100%);
+}
+
+.link:focus {
+  outline: none;
+  background-color: rgba(0, 0, 0, 0.1);
+}
