+++ src/constants/links.js
@@ -1,0 +1,9 @@
+const links = {
+  carbon: `http://www.carbondesignsystem.com/`,
+  github: `https://github.com/hellobrian`,
+  linkedin: `https://www.linkedin.com/in/hellobrian`,
+  blog: `https://medium.com/@brianhan`,
+  email: `mailto:thisisbrianhan@gmail.com`,
+}
+
+export default links
