--- src/pages/page-2.js
@@ -1,15 +1,0 @@
-import React from 'react'
-import { Link } from 'gatsby'
-import Layout from '../components/Layout'
-
-const SecondPage = () => (
-  <Layout>
-    <div>
-      <h1>Hi from the second page</h1>
-      <p>Welcome to page 2</p>
-      <Link to="/">Go back to the homepage</Link>
-    </div>
-  </Layout>
-)
-
-export default SecondPage
