+++ src/pages/pandas-and-bananas.md
@@ -1,0 +1,9 @@
+---
+title: 'Pandas and Bananas'
+date: '2017-08-21'
+---
+
+Do Pandas eat bananas? Check out this short video that shows that yes! pandas do
+seem to really enjoy bananas!
+
+<iframe width="560" height="315" src="https://www.youtube.com/embed/4SZl1r2O_bY" frameborder="0" allowfullscreen></iframe>
