+++ src/pages/sweet-pandas-eating-sweets.md
@@ -1,0 +1,10 @@
+---
+title: 'Sweet Pandas Eating Sweets'
+date: '2017-08-10'
+---
+
+Pandas are really sweet.
+
+Here's a video of a panda eating sweets.
+
+<iframe width="560" height="315" src="https://www.youtube.com/embed/4n0xNbfJLR8" frameborder="0" allowfullscreen></iframe>
